refactor: api 관련 리팩토링
- endPoint, constant 분리
- 쿠폰관련 api를 coupon 파일로 분리
