+++ src/apis/constants.ts
@@ -1,0 +1,3 @@
+export const API_BASE_URL = import.meta.env.VITE_API_URL as string;
+export const USER_ID = import.meta.env.VITE_USER_ID as string;
+export const USER_PASSWORD = import.meta.env.VITE_USER_PASSWORD as string;
