+++ src/apis/shoppingCart.ts
@@ -1,14 +1,13 @@
-import { CartItem, Coupon } from '@appTypes/shoppingCart';
+import { CartItem } from '@appTypes/shoppingCart';
 import { generateBasicToken } from '@utils/auth';
 
-const API_URL = import.meta.env.VITE_API_URL as string;
-const USER_ID = import.meta.env.VITE_USER_ID as string;
-const USER_PASSWORD = import.meta.env.VITE_USER_PASSWORD as string;
+import { USER_ID, USER_PASSWORD } from './constants';
+import endPoint from './endPoint';
 
 export async function fetchCartItems(): Promise<CartItem[]> {
   const token = generateBasicToken(USER_ID, USER_PASSWORD);
 
-  const response = await fetch(`${API_URL}/cart-items`, {
+  const response = await fetch(endPoint.cartItems, {
     method: 'GET',
     headers: { Authorization: token },
   });
@@ -27,7 +26,7 @@ export async function fetchCartItems(): Promise<CartItem[]> {
 export async function fetchCartItemCount(productId: number, quantity: number) {
   const token = generateBasicToken(USER_ID, USER_PASSWORD);
 
-  const response = await fetch(`${API_URL}/cart-items/${productId}`, {
+  const response = await fetch(endPoint.cartItem(productId), {
     method: 'PATCH',
     headers: {
       'Content-Type': 'application/json',
@@ -44,7 +43,7 @@ export async function fetchCartItemCount(productId: number, quantity: number) {
 export async function fetchDeleteCartItem(productId: number) {
   const token = generateBasicToken(USER_ID, USER_PASSWORD);
 
-  const response = await fetch(`${API_URL}/cart-items/${productId}`, {
+  const response = await fetch(endPoint.cartItem(productId), {
     method: 'DELETE',
     headers: {
       'Content-Type': 'application/json',
@@ -55,21 +54,4 @@ export async function fetchDeleteCartItem(productId: number) {
   if (!response.ok) {
     throw new Error('Failed to delete cart item');
   }
-}
-
-export async function fetchCouponList(): Promise<Map<string, Coupon>> {
-  const token = generateBasicToken(USER_ID, USER_PASSWORD);
-
-  const response = await fetch(`${API_URL}/coupons`, {
-    method: 'GET',
-    headers: { Authorization: token },
-  });
-
-  if (!response.ok) {
-    throw new Error('Failed to fetch products');
-  }
-
-  const data = await response.json();
-
-  return new Map(data.map((coupon: Coupon) => [coupon.code, coupon]));
 }
